# back/docs/swagger.yaml
basePath: /
definitions:
...
  request.GenerateDiagramRequest:
    type: object
  request.GenerateModelRequest:
    properties:
      modelName:
        example: MyModel
        type: string
      modelType:
        example: DEVS
        type: string
      previousModelsCode:
        example: Existing model code
        type: string
      userPrompt:
        example: Generate a model based on the previous code
        type: string
    required:
    - modelName
    - modelType
    - previousModelsCode
    - userPrompt
    type: object
...
  request.LoginRequest:
    properties:
      identity:
        type: string
      password:
        type: string
    required:
    - identity
    - password
    type: object
  request.LogoutRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
...
  request.RefreshRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  request.RegisterRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 50
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  response.Connection:
    properties:
      from:
        allOf:
        - $ref: '#/definitions/response.Endpoint'
        description: obligatoire
      to:
        allOf:
        - $ref: '#/definitions/response.Endpoint'
        description: obligatoire
    required:
    - from
    - to
    type: object
  response.DiagramResponse:
    properties:
      connections:
        description: obligatoire
        items:
          $ref: '#/definitions/response.Connection'
        type: array
      models:
        description: obligatoire
        items:
          $ref: '#/definitions/response.Model'
        type: array
    required:
    - connections
    - models
    type: object
  response.Endpoint:
    properties:
      model:
        description: obligatoire
        type: string
      port:
        description: obligatoire
        type: string
    required:
    - model
    - port
    type: object
  response.GeneratedModelResponse:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  response.LoginResponse:
    properties:
      accessToken:
        type: string
      email:
        type: string
      refreshToken:
        type: string
      username:
        type: string
    required:
    - accessToken
    - email
    - refreshToken
    - username
    type: object
  response.Model:
    properties:
      components:
        description: optionnel
        items:
          type: string
        type: array
      id:
        description: obligatoire
        type: string
      ports:
        allOf:
        - $ref: '#/definitions/response.Ports'
        description: optionnel
      type:
        allOf:
        - $ref: '#/definitions/response.ModelType'
        description: enum obligatoire
    required:
    - id
    - type
    type: object
...
  response.ModelType:
    enum:
    - atomic
    - coupled
    type: string
    x-enum-varnames:
    - ModelTypeAtomic
    - ModelTypeCoupled
  response.Ports:
    properties:
      in:
        items:
          type: string
        type: array
      out:
        items:
          type: string
        type: array
    type: object
  response.RefreshResponse:
    properties:
      accessToken:
        type: string
    required:
    - accessToken
    type: object
  response.RegisterResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/response.UserResponse'
    required:
    - accessToken
    - refreshToken
    - user
    type: object
  response.UserResponse:
    properties:
      email:
        type: string
      username:
        type: string
    required:
    - email
    - username
    type: object
host: localhost:3000
info:
  contact: {}
  description: This is the API documentation for Easy DEVS.
  title: Easy DEVS API
  version: "1.0"
paths:
  /ai/generate-diagram:
    post:
      consumes:
      - application/json
      description: Sends a prompt to OpenAI to generate a diagram in JSON format based
        on a strict schema.
      parameters:
      - description: Data required to generate a diagram
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.GenerateDiagramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Generated diagram
          schema:
            $ref: '#/definitions/response.DiagramResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: AI processing error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Generate a diagram
      tags:
      - AI
  /ai/generate-model:
    post:
      consumes:
      - application/json
      description: Sends a prompt to OpenAI to generate a DEVS model code.
      parameters:
      - description: Data required to generate a model
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.GenerateModelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Generated model code
          schema:
            $ref: '#/definitions/response.GeneratedModelResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: AI processing error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Generate a model
      tags:
      - AI
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user with the provided identity (email or username) and
        password.
      parameters:
      - description: Login Request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Log in a user
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logs out a user by invalidating the refresh token.
      parameters:
      - description: Logout Request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Log out a user
      tags:
      - Auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Returns the authenticated user's information based on the access
        token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get current user
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refreshes the access token using a valid refresh token.
      parameters:
      - description: Refresh Request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RefreshResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user with the provided username, email, and password.
      parameters:
      - description: Register Request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Auth
...
swagger: "2.0"